refactor: replace AddImage component with ImageUploadDrawer and add ImageGenDrawer for enhanced image handling
